refactor: tournament 디렉토리명 match-up으로 변경
