+++ frontend/src/pages/custom-game-list/page.js
@@ -242,7 +242,7 @@ export default function CustomGameList($container) {
     // 방 걸러보기
     const $1vs1FilterBtn = document.getElementById("1vs1-filter-btn");
     const $tournamentFilterBtn = document.getElementById(
-      "tournament-filter-btn",
+      "match-up-filter-btn",
     );
 
     click($1vs1FilterBtn, () => {
@@ -280,7 +280,7 @@ export default function CustomGameList($container) {
 
     // 게임 방 만들기 모달
     const $1vs1ModeBtn = document.getElementById("1vs1-toggle-btn");
-    const $tournamentModeBtn = document.getElementById("tournament-toggle-btn");
+    const $tournamentModeBtn = document.getElementById("match-up-toggle-btn");
     const $makeRoomBtn = document.getElementById("make-room-btn");
 
     click($1vs1ModeBtn, () => {
